requirements.txt
--extra-index-url https://download.pytorch.org/whl/cpu

# PyTorch (CPU only)
torch

# Reranking (CPU)
sentence-transformers
scikit-learn             # For BM25 and text processing dependencies

# BM25 for hybrid search
pinecone-text

# LangChain + Retrieval + OpenAI embedding/LLM
openai
langchain
langchain-core
langchain-openai

# Vector database
pinecone

# Redis for queue (Upstash/Redis)
redis
python-dotenv

# PDF and requests
PyMuPDF
requests
nltk                      # For punkt/stopwords, run download at build!

# Utilities
tqdm                      # Nice progress bar for batch jobs (optional, not required)
